pyproject.toml
[tool.poetry]
name = "adventofcode2024"
version = "0.1.0"
description = ""
authors = ["Gerben Plijter <gplijter@umaco.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
httpx = "^0.28.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
